Rename tools to common
