+++ app/src/main/java/com/hrb116/waterlogged/common/tokens/TokenManager.kt
@@ -1,9 +1,9 @@
-package com.hrb116.waterlogged.tools.tokens
+package com.hrb116.waterlogged.common.tokens
 
 import android.content.Context
 import android.util.Log
 import com.hrb116.waterlogged.BuildConfig
-import com.hrb116.waterlogged.tools.doPostRequest
+import com.hrb116.waterlogged.common.doPostRequest
 import kotlinx.coroutines.CancellationException
 import java.time.LocalDateTime
