+++ app/src/main/java/com/hrb116/waterlogged/tile/MainTileService.kt
@@ -8,15 +8,15 @@ import com.hrb116.waterlogged.R
 import com.hrb116.waterlogged.tile.addwater.addWaterLayout
 import com.hrb116.waterlogged.tile.addwater.waterLayout
 import com.hrb116.waterlogged.tile.login.loginLayout
-import com.hrb116.waterlogged.tools.tokens.getValue
-import com.hrb116.waterlogged.tools.tokens.isTokenExpired
-import com.hrb116.waterlogged.tools.postWater
-import com.hrb116.waterlogged.tools.tokens.refreshTokens
+import com.hrb116.waterlogged.common.tokens.getValue
+import com.hrb116.waterlogged.common.tokens.isTokenExpired
+import com.hrb116.waterlogged.common.postWater
+import com.hrb116.waterlogged.common.tokens.refreshTokens
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
 import com.google.android.horologist.tiles.SuspendingTileService
 import com.google.android.horologist.tiles.images.drawableResToImageResource
-import com.hrb116.waterlogged.tools.WaterContainers
-import com.hrb116.waterlogged.tools.tokens.Tokens
+import com.hrb116.waterlogged.common.WaterContainers
+import com.hrb116.waterlogged.common.tokens.Tokens
 
 private const val RESOURCES_VERSION = "0"
